fix (zookeeper): IHost instance is not being properly disposed when StopAsync is called before disposal
IHost instance is not being properly disposed when StopAsync is called before disposal, leaving background services running and causing tests to hang. Changed the entire approach in the IsZooKeeperAvailableAsync() method to create an instance of the ZooKeeperLeaderElectionProvider class and calling the health check method to check whether ZooKeeper is available.

Fixes #1