fix (postgres,sqlserver): IHost instance is not being properly disposed when StopAsync is called before disposal
Similar to the issue for ZooKeeper. IHost instance is not being properly disposed when StopAsync is called before disposal, leaving background services running and causing tests to hang. Changed the entire approach in the IsPostgreSqlAvailableAsync() and IsSqlServerAvailableAsync() methods to create an instance of the LeaderElextion provider classes and calling the health check method to check whether PostgreSQL and SqlServer are available.

Fixes #1